extract utilities
