+++ databases/catdat/scripts/deduce.ts
@@ -2,7 +2,7 @@ import { deduce_category_implications } from './deduce-category-implications'
 import { deduce_functor_implications } from './deduce-functor-implications'
 import { deduce_special_objects } from './deduce-special-objects'
 import { deduce_special_morphisms } from './deduce-special-morphisms'
-import { deduce_properties_for_entities } from './utils/deduction'
+import { deduce_properties_for_entities } from './deduce-entity-properties'
 
 deduce()
