+++ databases/catdat/scripts/utils/implications.ts
@@ -1,0 +1,58 @@
+import type { NormalizedImplication, PropertyMeta } from './deduction'
+
+function get_assumption_string(
+	implication: NormalizedImplication,
+	properties_dict: Record<string, PropertyMeta>,
+	conditional = false,
+): string {
+	const { assumptions } = implication
+
+	return Array.from(assumptions)
+		.map(
+			(assumption) =>
+				`${properties_dict[assumption][conditional ? 'conditional' : 'relation']} ${assumption}`,
+		)
+		.join(' and ')
+}
+
+function get_conclusion_string(
+	implication: NormalizedImplication,
+	properties_dict: Record<string, PropertyMeta>,
+	conditional = false,
+): string {
+	const { conclusion } = implication
+
+	return `${properties_dict[conclusion][conditional ? 'conditional' : 'relation']} ${conclusion}`
+}
+
+export function get_reason_string(
+	implication: NormalizedImplication,
+	properties_dict: Record<string, PropertyMeta>,
+	type: 'category' | 'functor',
+) {
+	const assumption_string = get_assumption_string(implication, properties_dict)
+	const conclusion_string = get_conclusion_string(implication, properties_dict)
+
+	const ref = `by <a href="/${type}-implication/${implication.id}">this result</a>`
+	return `Since it ${assumption_string}, it ${conclusion_string} (${ref}).`
+}
+
+export function get_contradiction_string(
+	implication: NormalizedImplication,
+	properties_dict: Record<string, PropertyMeta>,
+	property: string,
+	type: 'category' | 'functor',
+) {
+	const assumption_string = get_assumption_string(implication, properties_dict, true)
+	const conclusion_string = get_conclusion_string(implication, properties_dict, true)
+
+	const has_multiple_assumptions = implication.assumptions.size > 1
+
+	const ref = `by <a href="/${type}-implication/${implication.id}">this result</a>`
+
+	const contra = `Assume for contradiction that it ${properties_dict[property].relation} ${property}`
+
+	return has_multiple_assumptions
+		? `${contra}. Then it ${assumption_string}, so it ${conclusion_string} (${ref}) – contradiction.`
+		: `${contra}. Then it ${conclusion_string} (${ref}) – contradiction.`
+}
